LayoutTests/fast/dom/Geolocation/permission-denied-stops-watches-expected.txt

Tests that when Geolocation permission is denied, watches are stopped, as well as one-shots.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true
PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"

TEST COMPLETE
